Add files via upload
This project is functionally a library database that uses a .dll file in the back-end component which the front-end relies on to display the data. The two main variables that are stored / manipulated in the database are the books and customers – and both are shown in the web application in tables, along with their IDs, book author etc. **In Progress**